Fix cnrps
